Remove GlobalParams (#34)
This PR removes the GlobalParams type and moves the executor ID and app
version getters to the DBOS Executor type. The logic to calculate the
app version from code remains in a utility class to avoid cluttering up
DBOS Executor with logic that is only used in one place.

fixes #33

~Note, still two tests disabled which is why I opened this in draft~