Disable code in shutdown path (#52)
conductor repo tests need to be able to shutdown and relaunch DBOS.
However, the current DBOS shutdown path closes the DB, shuts down the
executor service and discards teh worklow registry.

This PR comments out that code to allow shutdown/relaunch to succeed.
Clearly, this is not a long term solution. Issue #51 tracks refactoring
the core objects and their dependencies to solve this issue.