Devhawk/refactoring-step-3 (#59)
this PR removes the static DBOS instance from DBOS class and makes DBOS
work purely in an instance model. Our expectation is that the core DBOS
implementation will be instance based, but then we can build static
wrapper and/or singleton Spring Boot beans from that core
implementation.

Note, one thing I realized in doing this work is that I had to inject
the dbos instance into the various workflow implementations that needed
to call things like `DBOS.send` This is not a good developer experience,
but will be addressed in a future PR.