Include DBOS instance in DBOS Context (#60)
while we focus on making DBOS work in Java as an instance, we need a
mechanism for WF methods to access capabilities on the associated dbos
instnace (durable sleep, send/recv, get & setEvent, etc). This PR adds a
DBOS instance field to DBOSContext + DBOSContext static methods for
retrieving that instance.

lifetime of internal classes like DBOSExecutor *will* change eventually
to be tied to the DBOS launch/shutdown lifetime, but for now we
centralize setting the dbos instance in context in DBOS Executor. What
we currently set the DBOS instance in 'DBOSExecutor.start', eventually
it will be injectect via ctor.

Also added DBOSContext statics for workflow id and in workflow, since
those are used in tests.

updated all tests to not manually manage the dbos instance in workflow
classes and to use DBOSContext statics instead