Uniform use of withers (#150)
Provide withers for ListWorkflowsInput, DBOSConfig, Queue, etc.

Configuration behavior also changed... hidden use of env variables,
which had been in `build()`, now only occurs if you call
`defaultsFromEnv`, which is convenient for tests but not referenced in
the README. We may continue to improve in this area, but it is better
than it was.

Also:
  Fix some issues with queue concurrency settings
  Test listWorkflows and listWorkflowSteps when called as steps.