add timeout sum type (#153)
Previously, you would specify a zero timeout to indicate a child
workflow had no timeout, even if the parent workflow had a timeout. This
was a poor developer experience.

This PR adds a new `Timeout` sum type that can be one of three values:
`Inherit`, `None`, or `Explicit` with `Expicit` including a timeout
`Duration`. This allows child worklfows to be started
`withTimeout(duration)` or `withNoTimeout()`. Workflow timeout defaults
to `Inherit` if not set (which is the same as `None` for top level
workflows)

fixes #88