Fix startWorkflow (#234)
`startWorkflow` requires information about the @Workflow executed in the
provided lambda. Previously, we added the start options to the context
and executed the workflow. The @Workflow info would end up in
invokeWorkflow where the behavior was different based on the presence of
start options. This approach creates thread pool exhaustion when
starting 1000s of workflows and creates significant complexity in
`DBOSContext` and `invokeWorkflow` to account for this difference in
execution.

This PR introduces a ThreadLocal hook method that `startWorkflow` can
set and `DBOSInvocationHandler` can call. This allows `startWorkflow` to
retrieve the class/instance/workflow names and the workflow argument
from the provided lambda without requiring a thread pool thread. Once we
have this information, we can call the private `executeWorkflow` method
directly. This simplifies the context significantly and eliminates the
need for `invokeWorkflow to do dual duty.

This PR also changes `executeWorkflow` to use
`CompletableFuture.supplyAsync` instead of `executorService.submit` to
run the workflow. This will allow us to leverage different executors
available on later versions of Java. However, this work will be in a
separate PR.

Note, the only changes to the tests are
* Consistent use of DBOSAwaitedWorkflowCancelledException instead of
CancellationException
* `parentTimeoutInheritedByChild` no longer requires retries

fixes #209
fixes #227